refactor(legal): consolidate legal-document writers, drop SaveChanges interceptor (nobodies-collective/Humans#751) (#1155)
- Merge AdminLegalDocumentService into LegalDocumentSyncService: one Application-layer
  service now owns all writes to legal_documents/document_versions, implementing both
  ILegalDocumentSyncService and IAdminLegalDocumentService.
- Service calls ILegalDocumentCacheInvalidator.InvalidateAll() directly after each
  successful repository write (create/update/archive/version-summary/sync-touch/
  version-add); interceptor deleted with its 4 AddInterceptors registrations
  (InfrastructureServiceCollectionExtensions, not Program.cs as the issue assumed).
- DI: IAdminLegalDocumentService resolves to the same keyed scoped instance as
  ILegalDocumentSyncService; decorator pattern unchanged.
- Interceptor tests replaced by service-level invalidation tests.

Note: the issue's target name "LegalDocumentService" was already taken by an unrelated
pre-existing service (public /Legal Statutes content provider, no DB access) - merged
class keeps the LegalDocumentSyncService name instead.
