+++ docs/sections/LegalAndConsent.md
@@ -156,7 +156,7 @@ Three controllers serve this section.
 ## Cross-Section Dependencies
 
 - **Profiles:** `IProfileService` — consent-check status lives on the profile (read by `ConsentService` for the review-detail view); `IProfileService.GetActiveApprovedUserIdsAsync` is the fan-out target list when `LegalDocumentSyncService` notifies on a new published / re-consent-required version. `ConsentService` does **not** call into Profile or Onboarding directly after a consent submit — the threshold check (`OnboardingService.SetConsentCheckPendingIfEligibleAsync`) is invoked by the controller (`ConsentController.Submit`, `OnboardingWidgetController`) as a peer call alongside `ConsentService.SubmitConsentAsync`.
-- **Teams:** `ITeamService` — `AdminLegalDocumentService` stitches team names in memory (replaces `.Include(d => d.Team)`); legal documents are team-scoped (Volunteers team = global).
+- **Teams:** `ITeamService` — `LegalDocumentSyncService` stitches team names in memory (replaces `.Include(d => d.Team)`); legal documents are team-scoped (Volunteers team = global).
 - **Notifications:** `INotificationService` (in-app fan-out from `LegalDocumentSyncService`) and `INotificationInboxService.ResolveBySourceAsync` (auto-resolve `AccessSuspended` notifications from `ConsentService` once all required consents are complete).
 - **Human Lifecycle:** `IHumanLifecycleService.RestoreConsentSuspensionAsync` — `ConsentService` lifts a consent suspension once all required consents are complete (alongside resolving the `AccessSuspended` notification). `ConsentService` no longer depends on `ISystemTeamSync` — after the name-only access switch, a consent submit does not provision system-team membership; the scheduled `SystemTeamSyncJob` reconciles Volunteers/Coordinators on name + consents.
 - **Governance:** `IMembershipCalculator.GetRequiredTeamIdsForUserAsync` / `HasAllRequiredConsentsAsync` — `ConsentService` resolves which teams' documents apply to a given user and whether all required consents are complete.
@@ -166,15 +166,15 @@ Three controllers serve this section.
 
 ## Architecture
 
-**Owning services:** `LegalDocumentService`, `AdminLegalDocumentService`, `LegalDocumentSyncService` (document-side), `ConsentService` (consent-side) — all in `Humans.Application.Services.Legal` / `Humans.Application.Services.Consent`.
+**Owning services:** `LegalDocumentService` (Statutes page), `LegalDocumentSyncService` (document-side — sole writer for `legal_documents`/`document_versions`, owning both the admin write surface `IAdminLegalDocumentService` and the GitHub-sync write surface `ILegalDocumentSyncService`; nobodies-collective/Humans#751), `ConsentService` (consent-side) — all in `Humans.Application.Services.Legal` / `Humans.Application.Services.Consent`.
 **Owned tables:** `legal_documents`, `document_versions`, `consent_records`
-**Status:** (A) Migrated — all four owning services live in `Humans.Application`; all table access routes through owning-section repositories. One cross-domain nav strip deferred (`LegalDocument.Team` — details below).
+**Status:** (A) Migrated — all three owning services live in `Humans.Application`; all table access routes through owning-section repositories. One cross-domain nav strip deferred (`LegalDocument.Team` — details below).
 
 - Services live in `Humans.Application.Services.Legal/` and `Humans.Application.Services.Consent/` and never import `Microsoft.EntityFrameworkCore`.
 - `ILegalDocumentRepository` (impl `LegalDocumentRepository` in `Humans.Infrastructure/Repositories/Legal/`) is the only code path that touches `legal_documents` and `document_versions` via `DbContext`.
 - `IConsentRepository` (impl `ConsentRepository` in `Humans.Infrastructure/Repositories/Consent/`) is the only code path that touches `consent_records` via `DbContext`. Exposes `AddAsync` and `GetXxxAsync` only — no `UpdateAsync`/`DeleteAsync`.
 - **Decorator decision (T-04, 2026-05-16)** — Two-layer cache landed:
-  - **Global** — `CachingLegalDocumentSyncService` (Singleton in Infrastructure) wraps `LegalDocumentSyncService` (inner, keyed Scoped). Holds the active+required document set as `LegalDocumentInfo[]` keyed by document id, plus a version-id → document-id index. Serves `GetActiveRequiredDocumentsForTeamsAsync`, `GetRequiredDocumentVersionsForTeamAsync`, `GetRequiredVersionsAsync`, `GetVersionByIdAsync` from cache. `LegalDocumentSaveChangesInterceptor` uses the dual-override pattern (snapshot affected entity state in `SavingChangesAsync`, consume in `SavedChangesAsync`, clear on failure) to flush the cache wholesale after any persisted write to `legal_documents` or `document_versions`. Eager warmup at startup: the decorator inherits `TrackedCache<Guid, LegalDocumentInfo>` with `warmOnStartup: true`, so its own `IHostedService.StartAsync` drives `WarmAllAsync` at boot (non-fatal); no separate warmup hosted service. Team names are stitched at warm time via `ITeamService.GetByIdsWithParentsAsync` so the cache build no longer walks the deprecated `LegalDocument.Team` cross-domain nav.
+  - **Global** — `CachingLegalDocumentSyncService` (Singleton in Infrastructure) wraps `LegalDocumentSyncService` (inner, keyed Scoped; also the sole writer, implementing both `ILegalDocumentSyncService` and `IAdminLegalDocumentService`). Holds the active+required document set as `LegalDocumentInfo[]` keyed by document id, plus a version-id → document-id index. Serves `GetActiveRequiredDocumentsForTeamsAsync`, `GetRequiredDocumentVersionsForTeamAsync`, `GetRequiredVersionsAsync`, `GetVersionByIdAsync` from cache. `LegalDocumentSyncService` calls `ILegalDocumentCacheInvalidator.InvalidateAll()` directly after each successful repository write (admin create/update/archive/version-summary, GitHub-sync version add/touch) — there is no SaveChanges interceptor (nobodies-collective/Humans#751). Eager warmup at startup: the decorator inherits `TrackedCache<Guid, LegalDocumentInfo>` with `warmOnStartup: true`, so its own `IHostedService.StartAsync` drives `WarmAllAsync` at boot (non-fatal); no separate warmup hosted service. Team names are stitched at warm time via `ITeamService.GetByIdsWithParentsAsync` so the cache build no longer walks the deprecated `LegalDocument.Team` cross-domain nav.
   - **Per-user** — `CachingConsentService` (Singleton in Infrastructure) wraps `ConsentService` (inner, keyed Scoped). Holds `UserConsentInfo` (the user's explicitly consented version-id set, with the merge source-id chain unioned at warm time) keyed by user id. Serves `GetConsentedVersionIdsAsync`, `GetConsentMapForUsersAsync`, and `GetRequiredConsentRowsForUserAsync` from cache. Lazy per-user warm. **Synchronous invalidation on `SubmitConsentAsync`**: the decorator override evicts the user (and any merged-source-id tombstones) inline before returning, so the controller's next-page consent-banner check observes the fresh state. `AccountMergeService.FoldAsync` evicts both source and target post-commit so the surviving target's cached set rebuilds against the new chain.
   - `LegalDocumentService` (Statutes page) keeps its own `IMemoryCache` for the GitHub-fetched anonymous statutes content (zero DB access, pure I/O cache; out of scope for T-04).
   - Architecture tests pin both the inner-service IMemoryCache-free constraint and the decorator's dual-interface shape: `ConsentArchitectureTests.{ConsentService_HasNoIMemoryCacheConstructorParameter, LegalDocumentSyncService_HasNoIMemoryCacheConstructorParameter, CachingConsentService_ImplementsBothServiceAndInvalidator, CachingLegalDocumentSyncService_ImplementsBothServiceAndInvalidator, CachingConsentService_DeclaresSubmitConsentAsync}`.
